+++ R/utils_model_parameters.R
@@ -347,12 +347,9 @@
     if (inherits(model, "mvord")) {
       rows <- params$Component != "correlation"
     } else if (is.null(params$Component)) {
-      # don't exponentiate dispersion
       rows <- seq_len(nrow(params))
-    } else if (inherits(model, c("clm", "clm2", "clmm"))) {
-      ## TODO: make sure we catch all ordinal models properly here
-      rows <- !tolower(params$Component) %in% c("location", "scale")
     } else {
+      # don't exponentiate dispersion
       rows <- !tolower(params$Component) %in% c("dispersion", "residual")
     }
     params[rows, columns] <- exp(params[rows, columns])